+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [7.25.0](https://github.com/Cap-go/capacitor-inappbrowser/compare/7.24.0...7.25.0) (2025-10-23)
+
+
+### Features
+
+* update Discord release webhook username to include repository name ([627edc8](https://github.com/Cap-go/capacitor-inappbrowser/commit/627edc87c56d949881065c82ec32fc4383f48608))
+
+
+### Bug Fixes
+
+* remove footer in discord ([65de5b5](https://github.com/Cap-go/capacitor-inappbrowser/commit/65de5b5b212e870382f04c52a906b320b70ebcd8))
+
+## [7.24.0](https://github.com/Cap-go/capacitor-inappbrowser/compare/7.23.0...7.24.0) (2025-10-23)
+
+
+### Features
+
+* add getPluginVersion method to InAppBrowserPlugin and update workflows for changelog generation ([2d98ecb](https://github.com/Cap-go/capacitor-inappbrowser/commit/2d98ecba55f44b5f16b5dba52b1f4165c1682a86))
+* remove default content message from Discord release webhook ([125e671](https://github.com/Cap-go/capacitor-inappbrowser/commit/125e67127bdc6c1e83d6c50face05632487e413f))
+
+## [7.23.0](https://github.com/Cap-go/capacitor-inappbrowser/compare/7.22.4...7.23.0) (2025-10-23)
+
+
+### Features
+
+* handle cases when custom scheme can't be opened ([be872dd](https://github.com/Cap-go/capacitor-inappbrowser/commit/be872dd88e472e1e46348c035c563a634366143a))
+* **ios:** support authorizedAppLinks within iOS universal link handling ([c04e0bc](https://github.com/Cap-go/capacitor-inappbrowser/commit/c04e0bcf3c6be6cc3b99f3b400b8ebcd0be80b76))
+
 ## [7.22.4](https://github.com/Cap-go/capacitor-inappbrowser/compare/7.22.3...7.22.4) (2025-10-16)
 
 ## [7.22.3](https://github.com/Cap-go/capacitor-inappbrowser/compare/7.22.2...7.22.3) (2025-10-15)